\name{swirl_testing-package}
\alias{swirl_testing-package}
\alias{swirl_testing}
\docType{package}
\title{
\packageTitle{swirl_testing}
}
\description{
\packageDescription{swirl_testing}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{swirl_testing}
\packageIndices{swirl_testing}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{swirl_testing}

Maintainer: \packageMaintainer{swirl_testing}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
